impl worker;

+++ src/worker/model/worker_node.rs
@@ -7,7 +7,7 @@ pub struct WorkerNode{
     pub id: Option<u64>,
     pub host_name: Option<String>,
     pub port: Option<String>,
-    pub worker_type: Option<i32>,
+    pub worker_type: Option<i8>,
     pub launch_date: Option<NaiveDateTime>,
     pub modified: Option<NaiveDateTime>,
     pub created: Option<NaiveDateTime>,
